Add server info
Uses the info and version commands to get various info about the docker
server.

My use-case for this is forwarding a lot of information to logstash
from various clusters and wanting to be able to track down the server
that the log entry is coming from quickly. Since fleet can deploy
containers to different servers based on rules and availability, getting
the hostname that the container lives on is vital.